Proposed fix for https://github.com/krakenjs/express-enrouten/issues/73
Instead of just trying to resolve the module, try to load it.

Since this block is already surrounded by a try/catch statement it's a safe
place to attempt requiring a module. I don't think it's wasted effort, since
the whole point of validating the modules is to try to load them later.

Also, some very minor refactoring of the extension removal code to make it more readable.