give some kind of error when require fails
In the event that a user is executing code before a module is exported there may be an error.
The try statement in `isFileModule` silences any errors that might occur at require time. 
This creates an annoying debugging issue where a route just won't be mounted and there won't be any explanation.

Also, there is no documentation about debugging and enabling the log messages.